requirements.txt
selenium
webdriver-manager
PyHamcrest
behave